% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexCondition.R
\name{ComplexCondition}
\alias{ComplexCondition}
\title{R6 class representing a complex condition.}
\description{
A complex condition is made up of multiple elementary conditions by forming their conjunction.
}
\details{
A complex condition is the conjunction of elementary conditions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conditions}}{the set elementary conditions}

\item{\code{covered}}{the set of matching objects - logical vector for all objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ComplexCondition$new()}}
\item \href{#method-complexCover}{\code{ComplexCondition$complexCover()}}
\item \href{#method-firstMetric}{\code{ComplexCondition$firstMetric()}}
\item \href{#method-secondMetric}{\code{ComplexCondition$secondMetric()}}
\item \href{#method-print}{\code{ComplexCondition$print()}}
\item \href{#method-clone}{\code{ComplexCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new ComplexCondition object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplexCondition$new(conditions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditions}}{the set of elementary conditions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-complexCover"></a>}}
\subsection{Method \code{complexCover()}}{
Method to calculate the set of objects matching this complex condition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplexCondition$complexCover()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the set of matching objects - logical vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-firstMetric"></a>}}
\subsection{Method \code{firstMetric()}}{
Method for calculating the first comparison metric used in the DOMLEM paper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplexCondition$firstMetric(G)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{the parameter of the metric - a vector describing a set of objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a numeric metric between 0 and 1. Higher is better.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-secondMetric"></a>}}
\subsection{Method \code{secondMetric()}}{
Method for calculating the second comparison metric used in the DOMLEM paper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplexCondition$secondMetric(G)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{the parameter of the metric - a vector describing a set of objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a numeric metric between 0 and 1. Higher is better.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplexCondition$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplexCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
