% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{convertMatrixToList}
\alias{convertMatrixToList}
\title{Function that filters a matrix and returns an object list}
\usage{
convertMatrixToList(m, objects)
}
\arguments{
\item{m}{a boolean matrix}

\item{objects}{the vector of object names}
}
\value{
a list of vectors
}
\description{
Function that filters a matrix and returns an object list
}
