% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominatingSet.R
\name{dominatedSet}
\alias{dominatedSet}
\title{This function calculates the dominated set of an object with respect to a criterion set.}
\usage{
dominatedSet(x, P, examples)
}
\arguments{
\item{x}{the object}

\item{P}{the criterion set}

\item{examples}{the examples in the decision table}
}
\value{
the set of objects that are dominated by object x
}
\description{
This function calculates the dominated set of an object with respect to a criterion set.
}
