% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceRelation.R
\name{dominatesOutrankOnly}
\alias{dominatesOutrankOnly}
\title{Function to determine whether x dominates y on criterion set P.
We assume q is of gain-type criterion.}
\usage{
dominatesOutrankOnly(x, y, P)
}
\arguments{
\item{x}{the left operand}

\item{y}{the right operand}

\item{P}{the set of criterions to test}
}
\value{
whether x dominates y on criterion set P
}
\description{
Function to determine whether x dominates y on criterion set P.
We assume q is of gain-type criterion.
}
