% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMLEM.R
\name{DOMLEM}
\alias{DOMLEM}
\title{R6 class representing the DOMLEM algorithm for mixed-type attributes.}
\description{
The DOMLEM algorithm extracts a minimal set decision rules from an information table.
}
\details{
This class stores methods used in extracting the decision rules.
This is an extension of the original DOMLEM algorithm to consider mixed-type attributes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{it}}{the information table to extract the decision rules from}

\item{\code{P}}{the P-attributes}

\item{\code{roughSets}}{the P-lower and P-upper approximations of the class unions}

\item{\code{rules}}{the extracted decision rules}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DOMLEM$new()}}
\item \href{#method-main}{\code{DOMLEM$main()}}
\item \href{#method-findRules}{\code{DOMLEM$findRules()}}
\item \href{#method-rulesCover}{\code{DOMLEM$rulesCover()}}
\item \href{#method-addMinimalRules}{\code{DOMLEM$addMinimalRules()}}
\item \href{#method-clone}{\code{DOMLEM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new DOMLEM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOMLEM$new(it, P)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{it}}{the infromation table to work on}

\item{\code{P}}{the set of attributes to create the rough sets}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-main"></a>}}
\subsection{Method \code{main()}}{
Main method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOMLEM$main()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findRules"></a>}}
\subsection{Method \code{findRules()}}{
Method to extract a given type of decision rules for a given class union approximation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOMLEM$findRules(approximation, P, t, ruleType)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approximation}}{the class union approximation to cover}

\item{\code{P}}{the set of attributes used for creating the approximations}

\item{\code{t}}{the t-parameter of the class union}

\item{\code{ruleType}}{the type of rule to extract}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the extracted decision rules
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rulesCover"></a>}}
\subsection{Method \code{rulesCover()}}{
Method to calculate the objects covered by a set of rules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOMLEM$rulesCover(it, rules)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{it}}{the infromation table to work on}

\item{\code{rules}}{the set of rules}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the covered objects - set of object names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMinimalRules"></a>}}
\subsection{Method \code{addMinimalRules()}}{
Method to add the minimal decision rules to the existing set of rules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOMLEM$addMinimalRules(existingRules, newRules)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{existingRules}}{the set of rules to be extended}

\item{\code{newRules}}{the set of rules to be added}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a set of decision rules
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOMLEM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
