% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionUnions.R
\name{downwardClassUnion}
\alias{downwardClassUnion}
\title{Function for calculating the downward class union}
\usage{
downwardClassUnion(examples, t)
}
\arguments{
\item{examples}{the examples in the decision table}

\item{t}{the decision class to compare to}
}
\value{
the set of objects in the downward class union
}
\description{
Function for calculating the downward class union
}
