% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementaryCondition.R
\name{ElementaryCondition}
\alias{ElementaryCondition}
\title{R6 class representing an elementary condition.}
\description{
An elementary condition is a filter condition on a single attribute.
}
\details{
This class stores the attribute, the filter value, and for dominance variables whether the value is a lower- or upper bound.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attribute}}{the name of the attribute we filter on}

\item{\code{value}}{the constant in the filter condition}

\item{\code{isLowerBound}}{for dominance attributes whether the filter condition is a lower-bound or upper-bound}

\item{\code{attributeType}}{the type of the attribute (ind, sim, or dom)}

\item{\code{alpha}}{for similarity attributes the alpha parameter}

\item{\code{beta}}{for similarity attributes the beta parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ElementaryCondition$new()}}
\item \href{#method-elementCover}{\code{ElementaryCondition$elementCover()}}
\item \href{#method-print}{\code{ElementaryCondition$print()}}
\item \href{#method-toString}{\code{ElementaryCondition$toString()}}
\item \href{#method-equals}{\code{ElementaryCondition$equals()}}
\item \href{#method-clone}{\code{ElementaryCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new ElementaryCondition object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementaryCondition$new(attribute, value, it, isLowerBound = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{the attribute we filter on}

\item{\code{value}}{the value we use in the filter condition}

\item{\code{it}}{the information table to use}

\item{\code{isLowerBound}}{[optional] for dominance operators, whether the filter condition is a lower- or upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementCover"></a>}}
\subsection{Method \code{elementCover()}}{
Method to calculate the set of objects matching this elementary condition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementaryCondition$elementCover(it)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{it}}{the information table to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the set of matching objects - set of object names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementaryCondition$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toString"></a>}}
\subsection{Method \code{toString()}}{
toString method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementaryCondition$toString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-equals"></a>}}
\subsection{Method \code{equals()}}{
Method to compare two ElementaryConditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementaryCondition$equals(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{the other elementary condition to compare to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a single boolean value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementaryCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
