% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{isSubsetArbitrary}
\alias{isSubsetArbitrary}
\title{Function to determine whether A is a subset of B - arbitrary sets.}
\usage{
isSubsetArbitrary(A, B)
}
\arguments{
\item{A}{the left operand - vector}

\item{B}{the right operand - vector}
}
\value{
the boolean value of the operation
}
\description{
Function to determine whether A is a subset of B - arbitrary sets.
}
