% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outrankingRelation.R
\name{outranks}
\alias{outranks}
\title{Function to determine whether x outranks y on criterion q.
We assume q is of gain-type criterion.}
\usage{
outranks(x, y, q)
}
\arguments{
\item{x}{the left operand}

\item{y}{the right operand}

\item{q}{the criterion to test}
}
\value{
whether x outranks y on criterion q
}
\description{
Function to determine whether x outranks y on criterion q.
We assume q is of gain-type criterion.
}
