% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityRelation.R
\name{similar}
\alias{similar}
\title{Function to determine whether x is similar to y on attribute q.
The alpha- and beta-parameters of the similarity metric should be provided as attributes of column q.}
\usage{
similar(x, y, q)
}
\arguments{
\item{x}{the left operand}

\item{y}{the right operand}

\item{q}{the attribute to test}
}
\value{
whether x is similar to y on attribute q
}
\description{
Function to determine whether x is similar to y on attribute q.
The alpha- and beta-parameters of the similarity metric should be provided as attributes of column q.
}
