% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionUnions.R
\name{upwardClassUnion}
\alias{upwardClassUnion}
\title{Function for calculating the upward class union}
\usage{
upwardClassUnion(examples, t)
}
\arguments{
\item{examples}{the examples in the decision table}

\item{t}{the decision class to compare to}
}
\value{
the set of objects in the upward class union
}
\description{
Function for calculating the upward class union
}
