% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceRelation.R
\name{dominates}
\alias{dominates}
\title{Function to determine whether x dominates y on the mixed attribute set P.}
\usage{
dominates(x, y, P)
}
\arguments{
\item{x}{the left operand}

\item{y}{the right operand}

\item{P}{the set of attributes to test}
}
\value{
whether x dominates y on attribute set P
}
\description{
Function to determine whether x dominates y on the mixed attribute set P.
}
