% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationTable.R
\name{InformationTable}
\alias{InformationTable}
\title{A class for storing all relevant information of an information table.}
\description{
It stores the decision table, the types of the attributes, and the alpha and beta values of the similarity attributes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InformationTable$new()}}
\item \href{#method-clone}{\code{InformationTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$new(decisionTable, types = NA, alpha = NA, beta = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
