% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationTable.R
\name{InformationTable}
\alias{InformationTable}
\title{R6 class representing an information table.}
\description{
An information table consists of the decision table and meta-data.
}
\details{
This class stores the decision table, and meta-data.
The decision table consists of an object identifier column, a decision column, and at least one additional attribute.
The meta-data consist of the attribute names, their types, and the alpha and beta values for similarity attributes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{decisionTable}}{the set of examples}

\item{\code{metaData}}{meta-data of the attributes, including their name and type, along with alpha and beta parameters for similarity variables}

\item{\code{objects}}{vector of object names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InformationTable$new()}}
\item \href{#method-isCompatible}{\code{InformationTable$isCompatible()}}
\item \href{#method-downwardClassUnion}{\code{InformationTable$downwardClassUnion()}}
\item \href{#method-upwardClassUnion}{\code{InformationTable$upwardClassUnion()}}
\item \href{#method-classUnions}{\code{InformationTable$classUnions()}}
\item \href{#method-partitionAttributes}{\code{InformationTable$partitionAttributes()}}
\item \href{#method-dominates}{\code{InformationTable$dominates()}}
\item \href{#method-dominatingAndDominatedSets}{\code{InformationTable$dominatingAndDominatedSets()}}
\item \href{#method-similar}{\code{InformationTable$similar()}}
\item \href{#method-dominatingSet}{\code{InformationTable$dominatingSet()}}
\item \href{#method-dominatedSet}{\code{InformationTable$dominatedSet()}}
\item \href{#method-upwardClassUnionUpperApproximation}{\code{InformationTable$upwardClassUnionUpperApproximation()}}
\item \href{#method-upwardClassUnionLowerApproximation}{\code{InformationTable$upwardClassUnionLowerApproximation()}}
\item \href{#method-downwardClassUnionUpperApproximation}{\code{InformationTable$downwardClassUnionUpperApproximation()}}
\item \href{#method-downwardClassUnionLowerApproximation}{\code{InformationTable$downwardClassUnionLowerApproximation()}}
\item \href{#method-roughSets}{\code{InformationTable$roughSets()}}
\item \href{#method-clone}{\code{InformationTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new information table object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$new(decisionTable, metaData = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decisionTable}}{data frame containing the decision examples}

\item{\code{metaData}}{data frame containing the meta-data of the attributes. This parameter is optional, and if not provided, we assume all dominance attributes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isCompatible"></a>}}
\subsection{Method \code{isCompatible()}}{
Method to determine whether another information table is compatible with this one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$isCompatible(it)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{it}}{the information table to compare to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downwardClassUnion"></a>}}
\subsection{Method \code{downwardClassUnion()}}{
Method for calculating the downward class union.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$downwardClassUnion(class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{the decision class to compare to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the set of objects in the downward class union
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upwardClassUnion"></a>}}
\subsection{Method \code{upwardClassUnion()}}{
Method for calculating the upward class union.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$upwardClassUnion(class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{the decision class to compare to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the set of objects in the upward class union
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classUnions"></a>}}
\subsection{Method \code{classUnions()}}{
Method for calculating all downward- and upward class unions at once.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$classUnions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a pair of matrices for both class unions, where each row represents a class, and each column represents an object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-partitionAttributes"></a>}}
\subsection{Method \code{partitionAttributes()}}{
Method that partitions attribute set P into into sets of the same attribute type.
Only types relevant for the dominance relation are considered (indiscernibility, similarity, and dominance).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$partitionAttributes(P)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{P}}{the set of attributes to partition - vector of attribute names}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of attribute sets
Function to determine whether x dominates y on the mixed attribute set P.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dominates"></a>}}
\subsection{Method \code{dominates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$dominates(x, y, P, compareSimilaritySwitched = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the left operand - object name}

\item{\code{y}}{the right operand - object name}

\item{\code{P}}{the set of attributes to test - vector of attribute names}

\item{\code{compareSimilaritySwitched}}{whether to test similarity with the parameters switched}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
whether x dominates y on attribute set P
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dominatingAndDominatedSets"></a>}}
\subsection{Method \code{dominatingAndDominatedSets()}}{
Method for calculating the P-dominated and P-dominating sets all at once.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$dominatingAndDominatedSets(P)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{P}}{the set of attributes to test - vector of attribute names
Method to determine whether x is similar to y on attribute q.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-similar"></a>}}
\subsection{Method \code{similar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$similar(x, y, q)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the left operand - a data frame}

\item{\code{y}}{the right operand - a data frame}

\item{\code{q}}{the attribute to test}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
whether x is similar to y on attribute q
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dominatingSet"></a>}}
\subsection{Method \code{dominatingSet()}}{
This method calculates the dominating set of an object with respect to a criterion set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$dominatingSet(x, P, compareSimilaritySwitched = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the object - object name}

\item{\code{P}}{the criterion set}

\item{\code{compareSimilaritySwitched}}{whether to test similarity with the parameters switched}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the set of objects that dominate object x
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dominatedSet"></a>}}
\subsection{Method \code{dominatedSet()}}{
This method calculates the dominated set of an object with respect to a criterion set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$dominatedSet(x, P, compareSimilaritySwitched = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the object - object name}

\item{\code{P}}{the criterion set}

\item{\code{compareSimilaritySwitched}}{whether to test similarity with the parameters switched}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the set of objects that are dominated by object x
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upwardClassUnionUpperApproximation"></a>}}
\subsection{Method \code{upwardClassUnionUpperApproximation()}}{
This method calculates the P-upper approximations of the upward class unions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$upwardClassUnionUpperApproximation(dominating_U)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dominating_U}}{the P-dominating sets (U) - matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the approximations for all classes in a boolean matrix from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upwardClassUnionLowerApproximation"></a>}}
\subsection{Method \code{upwardClassUnionLowerApproximation()}}{
This method calculates the P-lower approximations of the upward class unions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$upwardClassUnionLowerApproximation(downward_U)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{downward_U}}{the P-upper approximations of the downward class unions - matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the approximations for all classes in a boolean matrix from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downwardClassUnionUpperApproximation"></a>}}
\subsection{Method \code{downwardClassUnionUpperApproximation()}}{
This method calculates the P-upper approximations of the downward class unions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$downwardClassUnionUpperApproximation(dominated_U)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dominated_U}}{the P-dominated sets (U) - matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the approximations for all classes in a boolean matrix from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downwardClassUnionLowerApproximation"></a>}}
\subsection{Method \code{downwardClassUnionLowerApproximation()}}{
This method calculates the P-lower approximations of the downward class unions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$downwardClassUnionLowerApproximation(upward_U)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{upward_U}}{the P-upper approximations of the upward class unions - matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the approximations for all classes in a boolean matrix from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-roughSets"></a>}}
\subsection{Method \code{roughSets()}}{
This method calculates the P-lower and P-upper approximations of class unions and boundary regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$roughSets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named list of the approximations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InformationTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
